luamodulemanager20, separete file classes interfaces, some clean ups
